techDebt(buRefactorTraining): Move bulkUpload validate utils into utils folder
