+++ server/routes/establishments/bulkUpload/validate/training/validateTraining.js
@@ -2,7 +2,7 @@ const { addDobTrainingMismatchError, trainingCompletedBeforeAgeFourteen } = requ
 
 const { validateTrainingCsv } = require('./validateTrainingCsv');
 const { establishmentNotFoundInFile, addNoEstablishmentError } = require('../shared/uncheckedEstablishment');
-const { createWorkerKey, deleteRecord } = require('../shared/utils');
+const { createWorkerKey, deleteRecord } = require('../../../../../utils/bulkUpload/validateUtils.js');
 const { addNoWorkerError, workerNotFoundInFile } = require('./uncheckedWorker');
 
 exports.validateTraining = async (training, myAPIWorkers, workersKeyed, allWorkersByKey, allEstablishmentsByKey) => {
